Genericons 4.0: Move to grunt and NPM for building.
This is a big change. It's dropping backwards compatability in favor of simplicity. It also moves to npm and grunt to make it easier to build your own versions, or add to the set. There are no social logos anymore, there are better separate sets for that.

By moving to NPM and breaking with the past, we hope to encourage and streamline contributions.