Add svg4everybody for better IE support.
Needs a bit more testing, but this extends IE support when using the
external method of insertion. This means it should work in IE8-10, whereas
before only the internal method worked for those versions of IE.

For simplicity it may be sensible to remove the internal option
altogether, so long as svg4everybody provides sufficient support.
Support for IE8 and below may require polyfilling.